add comments + refactor duplicate code
