+++ backend/src/util/scraper/database/scraper.db.facade.ts
@@ -5,6 +5,11 @@ import { ScraperAttributesDbService } from "./scraper.attributes.db.service";
 import { ScraperBlogDbService } from "./scraper.blog.db.service";
 import { ScraperMediaDbService } from "./scraper.media.db.service";
 
+/**
+ * The Manager is an object that acts as an entry point for all the other services.
+ * It allows you to only import this object and reach all the others through this.
+ * Main purpose is to simplify imports elsewhere.
+ */
 @Injectable()
 export class ScraperDbManager {
   constructor(
